chore: move common code to util.go
